fix: export bridge types
