+++ packages/sdk-router/src/types/index.ts
@@ -1,2 +1,3 @@
+export * from './bridge'
 export * from './intent'
 export * from './misc'
