+++ packages/sdk-router/src/types/misc.ts
@@ -1,1 +1,2 @@
+export { FeeConfig, Query } from '../module'
 export { Slippage } from '../swap'
